requirements.txt
streamlit==1.40.0
streamlit-authenticator==0.4.1
pandas==2.1.3
numpy==1.25.0
plotly==5.16.0
PyYAML==6.0
xlsxwriter==3.2.0
transformers==4.47.1
llama_stack==0.0.63